Change type of Saturation and Value to float. Add test. Fix methods

+++ inc/hsv.h
@@ -3,8 +3,8 @@
 struct HSV
 {
 	unsigned short int Hue;
-	unsigned short int Saturation;
-	unsigned short int Value;
+	float Saturation;
+	float Value;
 };
 
 inline bool operator ==(const HSV & lhs, const HSV & rhs)
